feat: useShop으로 기능 분리
