+++ src/hooks/useShop.tsx
@@ -1,5 +1,5 @@
-import { useQuery } from "@tanstack/react-query";
-import { loadSelectBundle } from "@/api/shop";
+import { useInfiniteQuery, useQuery } from "@tanstack/react-query";
+import { loadBundle, loadSelectBundle } from "@/api/shop";
 import QUERYKEYS from "@/constants/querykey";
 
 export default function useShop() {
@@ -11,8 +11,36 @@ export default function useShop() {
     );
     return { data, refetch };
   };
+  const {
+    data: infiniteData,
+    fetchNextPage,
+    hasNextPage,
+    isFetchingNextPage,
+  } = useInfiniteQuery(
+    [QUERYKEYS.LOAD_BUNDLE], // 쿼리 키
+    ({ pageParam = 0 }) => loadBundle({ page: pageParam, size: 1 }),
+    {
+      getNextPageParam: (lastPage) => {
+        const nextPage = lastPage.data.pageable.pageNumber + 1;
+        return nextPage < lastPage.data.totalPages ? nextPage : undefined;
+      },
+    },
+  );
+
+  const handleScroll = () => {
+    if (
+      window.innerHeight + window.scrollY >= document.body.offsetHeight - 100 &&
+      hasNextPage &&
+      !isFetchingNextPage
+    ) {
+      fetchNextPage();
+    }
+  };
 
   return {
     useLoadSelectBundle,
+    infiniteData,
+    handleScroll,
+    isFetchingNextPage,
   };
 }
